scx_rusty: Move stats structs to stats.rs and rename for consistency
